requirements.txt
streamlit
tensorflow
opencv-python
numpy
matplotlib
seaborn
Pillow
